Move decompilation test implementation to separate class
